refactor(speech): lift voice-loop backend builder into primer-speech
Move `ChannelStt`, `run_audio_thread`, and the cpal mic/speaker + VAD +
whisper STT + piper TTS + on_audio + drain-hook construction out of
primer-cli/src/speech_loop into a new
primer_speech::voice_loop::backends module so the GUI's voice mode can
reuse the same builder without duplicating CLI-private code.

primer-cli now constructs `LocalBackends` via the shared helper and
drives `run_loop_borrowed` over its fields. GUI's voice::backends is a
thin wrapper that calls the same builder. Every consumable field on
`LocalBackends` is `Option<T>` so the caller `.take()`s each one once
when handing it to `run_loop[_borrowed]`. `LocalBackends::shutdown()`
joins the audio thread; the Drop impl defends against missed calls.

All 76 primer-speech tests pass; primer-cli still builds with and
without --features speech.
