+++ src/crates/primer-gui/src/voice/mod.rs
@@ -3,6 +3,8 @@
 //! - [`observer`] — `TauriEventObserver` impl that emits `primer://voice/*`
 //!   events for the frontend.
 //! - [`assets`] — voice-asset path resolver + `AssetMissing` shape.
+//! - [`backends`] — thin wrapper around
+//!   `primer_speech::voice_loop::build_local_backends` (cpal + VAD + STT + TTS).
 //! - [`responder`] — `ArcDmResponder`: a `Responder` impl that locks a
 //!   shared `Arc<Mutex<DialogueManager>>` per turn.
 //!
@@ -12,6 +14,8 @@
 
 #[cfg(feature = "speech")]
 pub mod assets;
+#[cfg(feature = "speech")]
+pub mod backends;
 #[cfg(feature = "speech")]
 pub mod observer;
 #[cfg(feature = "speech")]
