+++ src/crates/primer-speech/src/voice_loop/mod.rs
@@ -17,9 +17,15 @@ pub mod locale_defaults;
 pub mod observer;
 pub mod state_machine;
 
+#[cfg(all(feature = "silero", feature = "whisper", feature = "piper", feature = "cpal"))]
+pub mod backends;
+
 pub use locale_defaults::{voice_default_for, LocaleDefault, LOCALE_DEFAULTS};
 pub use observer::{ExitReason, LoopObserver, TurnCompletePayload, VoiceState};
 pub use state_machine::{
     run_loop, run_loop_borrowed, DrainHook, LoopBackends, LoopConfig, LoopHandle, Responder,
     VoiceLoopError, VAD_EVENT_CHANNEL_CAPACITY,
 };
+
+#[cfg(all(feature = "silero", feature = "whisper", feature = "piper", feature = "cpal"))]
+pub use backends::{build_local_backends, ChannelStt, LocalBackends};
